+++ lib/td_auth_web/views/password_view.ex
@@ -1,0 +1,16 @@
+defmodule TdAuthWeb.PasswordView do
+  use TdAuthWeb, :view
+  alias TdAuthWeb.UserView
+
+  def render("show.json", %{user: user} = assigns) do
+    %{
+      data:
+        render_one(
+          user,
+          UserView,
+          "user.json",
+          Map.drop(assigns, [:user])
+        )
+    }
+  end
+end