+++ test/support/factory.ex
@@ -4,13 +4,15 @@ defmodule TdAuth.Factory do
   """
 
   use ExMachina.Ecto, repo: TdAuth.Repo
+  @password "secret hash"
 
   def user_factory do
+    secret =  Bcrypt.add_hash(@password)
     %TdAuth.Accounts.User{
       user_name: sequence("username"),
       email: sequence(:email, &"username#{&1}@example.com"),
       full_name: sequence("fullname"),
-      password_hash: "secret hash",
+      password_hash: secret.password_hash,
       groups: []
     }
   end